+++ Aplicacao/backend/PDV.Domain/Entities/ProdutoEntity.cs
@@ -10,8 +10,8 @@ namespace PDV.Domain.Entities
         public string Tamanho { get; set; } = string.Empty;
         public string Cor { get; set; } = string.Empty;
         public int Quantidade { get; set; }
-        public decimal PrecoCusto { get; set; }
-        public decimal PrecoVenda { get; set; }
+        public double PrecoCusto { get; set; }
+        public double PrecoVenda { get; set; }
         public DateTime CriadoEm { get; set; }
 
         public IEnumerable<ItemVendaEntity> ItensVenda { get; set; } = new List<ItemVendaEntity>();
